+++ test/drawbox.test.ts
@@ -23,7 +23,7 @@ describe('DrawBox module', () => {
         const a = r.draw(99, 40);
         const b = !((a[1] === 1) && (a[2] === 2) && (a[3] === 3) && (a[4] === 4))
         expect(b).toBe(true);
-        expect(() => { new DrawBox(200); }).toThrow(Error);
+        expect(() => { new DrawBox(0); }).toThrow(Error);
     });